backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2==2.9.10
httpx==0.27.2
python-multipart==0.0.9
boto3==1.34.162
pillow==10.4.0
openai==1.44.1
requests==2.32.3